requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
altair==5.3.0
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
blinker==1.8.2
braceexpand==0.1.7
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
contourpy==1.2.1
cycler==0.12.1
cytoolz==0.12.3
datasets==2.20.0
decorator==5.1.1
dill==0.3.8
docopt==0.6.2
editdistance==0.8.1
einops==0.8.0
elastic-transport==8.13.1
elasticsearch==8.14.0
exceptiongroup==1.2.1
executing==2.0.1
fiddle==0.3.0
filelock==3.15.1
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.20.3
grpcio==1.64.1
huggingface-hub==0.23.4
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.7
inflect==7.2.1
intervaltree==3.1.0
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.4
jiwer==3.0.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
julius==0.2.7
kiwisolver==1.4.5
lazy_loader==0.4
lhotse==1.24.1
libcst==1.4.0
librosa==0.10.2.post1
lightning==2.3.0
lightning-utilities==0.11.2
lilcom==1.7
llvmlite==0.43.0
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
nemo_toolkit @ git+https://github.com/NVIDIA/NeMo.git@501f0dfc76886fda7f95e934de39fd8275628e2a
networkx==3.3
numba==0.60.0
numpy==2.0.0
omegaconf==2.3.0
onnx==1.16.1
openai-whisper==20231117
optuna==3.6.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.2
primePy==1.3
prompt_toolkit==3.0.47
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.audio==3.3.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycparser==2.22
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.0
pytorch-metric-learning==2.5.0
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
speechbrain==1.0.0
SQLAlchemy==2.0.31
stack-data==0.6.3
streamlit==1.36.0
sympy==1.12.1
tabulate==0.9.0
tenacity==8.4.2
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.3.1
torchmetrics==1.4.0.post0
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
typeguard==4.3.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
webdataset==0.2.86
Werkzeug==3.0.3
wget==3.2
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
